backend/.ipynb_checkpoints/requirements-checkpoint.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
pillow==11.1.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
uvicorn==0.34.0
